# src/interface/api/routes/imports.py
"""Import trigger and checkpoint status endpoints.

Each import endpoint pre-generates an operation_id, registers an SSE queue,
launches the import as a background task, and immediately returns the
operation_id so the client can subscribe to progress via SSE.
"""

import os
from pathlib import Path
import tempfile

from fastapi import APIRouter, Depends, HTTPException, UploadFile

from src.config import get_logger
from src.config.constants import BusinessLimits
from src.infrastructure.connectors.spotify.auth import RECENTLY_PLAYED_SCOPE
from src.interface.api.deps import (
    get_current_user_id,
    require_connector_connected,
    require_connector_scopes,
)
from src.interface.api.schemas.imports import (
    CheckpointStatusSchema,
    ExportLastfmLikesRequest,
    ImportLastfmHistoryRequest,
    ImportSpotifyLikesRequest,
    ImportSpotifyRecentRequest,
    OperationStartedResponse,
)
from src.interface.api.services.progress import OperationBoundEmitter
from src.interface.api.services.sse_operations import launch_sse_operation

logger = get_logger(__name__).bind(service="imports_api")

router = APIRouter(prefix="/imports", tags=["imports"])


# ---------------------------------------------------------------------------
# Import endpoints
# ---------------------------------------------------------------------------


@router.post("/lastfm/history")
async def import_lastfm_history(
    body: ImportLastfmHistoryRequest,
    user_id: str = Depends(get_current_user_id),
    _connected: None = Depends(require_connector_connected("lastfm")),
) -> OperationStartedResponse:
    """Trigger a Last.fm listening history import."""

    async def _import(emitter: OperationBoundEmitter) -> object:
        from src.application.use_cases.import_play_history import run_import

        return await run_import(
            user_id=user_id,
            service="lastfm",
            mode=body.mode,
            limit=body.limit,
            from_date=body.from_date,
            to_date=body.to_date,
            progress_emitter=emitter,
        )

    return await launch_sse_operation(
        user_id=user_id,
        operation_type="import_lastfm_history",
        coro_factory=_import,
    )


@router.post("/spotify/recent")
async def import_spotify_recent(
    body: ImportSpotifyRecentRequest,
    user_id: str = Depends(get_current_user_id),
    _scoped: None = Depends(
        require_connector_scopes("spotify", {RECENTLY_PLAYED_SCOPE})
    ),
) -> OperationStartedResponse:
    """Poll Spotify's recently-played API for new plays.

    Scope-gated rather than merely connection-gated: a grant minted before
    v0.10.1 still works for likes and playlists, so the generic connected check
    would let it through and fail later inside the operation.
    """

    async def _import(emitter: OperationBoundEmitter) -> object:
        from src.application.use_cases.import_play_history import run_import

        return await run_import(
            user_id=user_id,
            service="spotify",
            mode="recent",
            limit=body.limit,
            progress_emitter=emitter,
            force=body.force,
        )

    return await launch_sse_operation(
        user_id=user_id,
        operation_type="import_spotify_recent",
        coro_factory=_import,
    )


@router.post("/spotify/likes")
async def import_spotify_likes(
    body: ImportSpotifyLikesRequest,
    user_id: str = Depends(get_current_user_id),
    _connected: None = Depends(require_connector_connected("spotify")),
) -> OperationStartedResponse:
    """Trigger a Spotify liked tracks import."""

    async def _import(emitter: OperationBoundEmitter) -> object:
        from src.application.use_cases.sync_likes import run_spotify_likes_import

        return await run_spotify_likes_import(
            user_id=user_id,
            limit=body.limit,
            max_imports=body.max_imports,
            force=body.force,
            progress_emitter=emitter,
        )

    return await launch_sse_operation(
        user_id=user_id,
        operation_type="import_spotify_likes",
        coro_factory=_import,
    )


@router.post("/lastfm/likes")
async def export_lastfm_likes(
    body: ExportLastfmLikesRequest,
    user_id: str = Depends(get_current_user_id),
    _connected: None = Depends(require_connector_connected("lastfm")),
) -> OperationStartedResponse:
    """Trigger a Last.fm likes export (love tracks on Last.fm)."""

    async def _export(emitter: OperationBoundEmitter) -> object:
        from src.application.use_cases.sync_likes import run_lastfm_likes_export

        return await run_lastfm_likes_export(
            user_id=user_id,
            batch_size=body.batch_size,
            max_exports=body.max_exports,
            progress_emitter=emitter,
        )

    return await launch_sse_operation(
        user_id=user_id,
        operation_type="export_lastfm_likes",
        coro_factory=_export,
    )


async def _stream_upload_to_fd(file: UploadFile, fd: int) -> bool:
    """Stream an upload to ``fd`` in 64KB chunks; return True if size limit exceeded.

    Stops writing as soon as the byte count passes ``MAX_UPLOAD_BYTES`` so an
    oversized upload never lands on disk in full.
    """
    bytes_written = 0
    while chunk := await file.read(64 * 1024):
        bytes_written += len(chunk)
        if bytes_written > BusinessLimits.MAX_UPLOAD_BYTES:
            return True
        os.write(fd, chunk)
    return False


@router.post("/spotify/history")
async def import_spotify_history(
    file: UploadFile,
    user_id: str = Depends(get_current_user_id),
) -> OperationStartedResponse:
    """Import Spotify listening history from a GDPR data export JSON file."""
    if file.size and file.size > BusinessLimits.MAX_UPLOAD_BYTES:
        raise HTTPException(
            status_code=413,
            detail=f"File too large ({file.size} bytes). Maximum is {BusinessLimits.MAX_UPLOAD_BYTES} bytes.",
        )

    # Save uploaded file to temp location with server-side byte counting.
    # Streaming 64KB chunks avoids loading the entire file into memory and
    # enforces the size limit regardless of what Content-Length claims.
    # Using os.* instead of pathlib for async-safe file I/O (ASYNC240).
    fd, temp_name = tempfile.mkstemp(suffix=".json")
    try:
        oversized = await _stream_upload_to_fd(file, fd)
    except BaseException:
        os.close(fd)
        os.unlink(temp_name)  # noqa: PTH108 — os.unlink is async-safe, pathlib is not (ASYNC240)
        raise
    else:
        os.close(fd)

    if oversized:
        os.unlink(temp_name)  # noqa: PTH108 — os.unlink is async-safe, pathlib is not (ASYNC240)
        raise HTTPException(
            status_code=413,
            detail=f"File too large (>{BusinessLimits.MAX_UPLOAD_BYTES} bytes). Maximum is {BusinessLimits.MAX_UPLOAD_BYTES} bytes.",
        )

    temp_path = Path(temp_name)

    async def _import(emitter: OperationBoundEmitter) -> object:
        from src.application.use_cases.import_play_history import run_import

        try:
            return await run_import(
                user_id=user_id,
                service="spotify",
                mode="file",
                file_path=temp_path,
                progress_emitter=emitter,
            )
        finally:
            os.unlink(temp_name)  # noqa: PTH108 — os.unlink is async-safe, pathlib is not (ASYNC240)

    return await launch_sse_operation(
        user_id=user_id,
        operation_type="import_spotify_history",
        coro_factory=_import,
    )


# ---------------------------------------------------------------------------
# Checkpoint status
# ---------------------------------------------------------------------------


@router.get("/checkpoints")
async def get_checkpoints(
    user_id: str = Depends(get_current_user_id),
) -> list[CheckpointStatusSchema]:
    """Get sync checkpoint status for all known service/entity combinations."""
    from src.application.use_cases.sync_likes import get_all_checkpoint_statuses

    statuses = await get_all_checkpoint_statuses(user_id=user_id)
    return [
        CheckpointStatusSchema(
            service=s.service,
            entity_type=s.entity_type,
            last_sync_timestamp=s.last_sync_timestamp,
            has_previous_sync=s.has_previous_sync,
            local_count=s.local_count,
            remote_total=s.remote_total,
        )
        for s in statuses
    ]