# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mixd",
    "description": "Personal music metadata hub",
    "version": "0.10.0"
  },
  "paths": {
...
    "/api/v1/imports/lastfm/history": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Import Lastfm History",
        "description": "Trigger a Last.fm listening history import.",
        "operationId": "import_lastfm_history_api_v1_imports_lastfm_history_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportLastfmHistoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStartedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/spotify/recent": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Import Spotify Recent",
        "description": "Poll Spotify's recently-played API for new plays.\n\nScope-gated rather than merely connection-gated: a grant minted before\nv0.10.1 still works for likes and playlists, so the generic connected check\nwould let it through and fail later inside the operation.",
        "operationId": "import_spotify_recent_api_v1_imports_spotify_recent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportSpotifyRecentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStartedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/spotify/likes": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Import Spotify Likes",
        "description": "Trigger a Spotify liked tracks import.",
        "operationId": "import_spotify_likes_api_v1_imports_spotify_likes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportSpotifyLikesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStartedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/lastfm/likes": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Export Lastfm Likes",
        "description": "Trigger a Last.fm likes export (love tracks on Last.fm).",
        "operationId": "export_lastfm_likes_api_v1_imports_lastfm_likes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportLastfmLikesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStartedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/spotify/history": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Import Spotify History",
        "description": "Import Spotify listening history from a GDPR data export JSON file.",
        "operationId": "import_spotify_history_api_v1_imports_spotify_history_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_spotify_history_api_v1_imports_spotify_history_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStartedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_spotify_history_api_v1_imports_spotify_history_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_spotify_history_api_v1_imports_spotify_history_post"
      },
...
      "ExportLastfmLikesRequest": {
        "properties": {
          "batch_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Size"
          },
          "max_exports": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Exports"
          }
        },
        "type": "object",
        "title": "ExportLastfmLikesRequest",
        "description": "Request body for triggering a Last.fm likes export."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportLastfmHistoryRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "recent",
              "incremental",
              "full"
            ],
            "title": "Mode",
            "default": "incremental"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "from_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Date"
          },
          "to_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Date"
          }
        },
        "type": "object",
        "title": "ImportLastfmHistoryRequest",
        "description": "Request body for triggering a Last.fm history import."
      },
      "ImportSpotifyLikesRequest": {
        "properties": {
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "max_imports": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Imports"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "title": "ImportSpotifyLikesRequest",
        "description": "Request body for triggering a Spotify likes import."
      },
      "ImportSpotifyRecentRequest": {
        "properties": {
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "title": "ImportSpotifyRecentRequest",
        "description": "Request body for polling Spotify's recently-played API.\n\nNo mode field: the stored cursor makes every poll incremental. ``limit`` is\nclamped to the endpoint's 50-play ceiling downstream. ``force`` ignores the\ncursor for one poll and re-reads the whole retained window \u2014 a recovery\nlever for a cursor that ran ahead of what was actually stored; re-ingesting\nseen plays is harmless (the ledger conflict-skips them)."
      },
...
      "OperationStartedResponse": {
        "properties": {
          "operation_id": {
            "type": "string",
            "title": "Operation Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "operation_id"
        ],
        "title": "OperationStartedResponse",
        "description": "Returned immediately when a long-running operation is launched.\n\n``run_id`` is the persistent ``OperationRun`` audit-log row id, present\nwhen the route writes one via the seam-level recorder. The frontend\ncaches it alongside ``operation_id`` so the post-run toast can deep-\nlink to ``/settings/imports?run=<run_id>`` without a second round trip."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}